Reduce the size of newblocks by 50%
THIS COMMIT IS PART OF OUR BRANDING for bnb100

Remove the closure library from being included in build/war. It isn't
needed.  The one file (so far) that is (a css file) is explicitly
copied over. By doing so we reduce the file count (greatly) in a
deployment and reduce the size of deployed appinventor by about half.

We also compile both GWT and Blockly to reduce the size of what has to
be loaded into a client's browser.

Change-Id: I0e88a8e22313bf99a02b445a8fda69044d3042d2