New Welcome System.
Supply a Splash Screen whenever App Inventor is started. Note: This
feature is controlled by the AppInventorFeatures.showSplashScreen()
function which defaults to false. You will have to set it to true to
see the effect of this change. The splash screen that is displayed
is a "plug" value, branding should supply the actual iframe to use.

Change-Id: I6be0baa0ea8ac41276d602af5728ff41e1045bb5