BRANDING: Reorder Companion compatibility checks
We normally first verify the companion package name and then if it is
the proper version. This is the “correct” way to do things as version
numbers for different App Inventor systems may not be congruent in any
rational way.

However the current MIT App Inventor setup package ships with a very old
companion that doesn’t provide the package name. This causes the
companion package name check to fail. So in this commit, which is not
intended for the “master” branch, we reverse the order of the
checks. This way when people install our setup package the correct
behavior will be experienced, namely that the version check will fail
and an auto-update will be offered.

Change-Id: I1f4c97b628ed2f47f5dbc8db3db60dcc6ab3bb54