Fix logic bug in ReplMgr when no assets are present
Commit 72e0ca15 added a callback to handle loading of assets to the
Companion. However, if the Companion is connected to a project without
assets, the callback to switch from ASSET state to CONNECTED state
will never fire. This commit checks whether the asset map has 0
elements, in which case it will synchronously fire the callback so
that projects without assets will immediately transition to CONNECTED
state.

Change-Id: Iba12232966752ed9cb61af0f97aa16ebcffeed4a