Defer gallery loading until user chooses to open it
Daily we have 750,000+ requests for resources via /ode/gallery. These
requests occur even if the user never actually accesses the gallery
since they are done at initialization of Ode. This commit defers the
initialization of the gallery UI so that it is only generated once the
user accesses the gallery.

Change-Id: Ic16ae5f45c8f79570d763408f5958170fa332ae1