Fix block dragging issues from the drawer
This commit addresses two issues with the blockly update:

1. Updates the flyout implementation in Blockly core to store
information about the last created block so that if many create
operations are fired only the first will be applied. This should
prevent a reported issue where many blocks are created at exactly the
same location when dragging from the flyout.

2. Change the z-index of ode-DialogBox to be greater than Blockly's
drag surface z-index so that dialogs that appear during drag
operations will not be occluded by the dialog. Prior to this change,
the block would be above the dialog and prevent the user from being
able to click any button to dismiss the dialog.

Change-Id: I07f19b971860f72ec512dfd9cb39bcb5207f3987