Add CompanionUpgradeHelper to ReplMgr
Emulator upgrades broke with 2.45 because the new companion is so
large it triggers an OutOfMemoryError in the emulator. This change
adds an app called CompanionUpgradeHelper. Its sole purpose is to
download the Companion directly to disk within the emulator. It then
runs the package installer to upgrade the Companion. In order to be
launched by the web interface, it has the same package and version
number as 2.45 Companion and replaces itself with the real companion
on success.

Change-Id: Id83460366b029ac2a4a9877be407013d202c25ec