Reset min size of button with images
If someone sets an image on a button in a non-Classic theme, the
minimum width/height settings in the theme will force the button to
sometimes resize the image arbitrarily to fit the minimum constraints
rather than allowing the button to assume the shape of the image like
it would in the Classic theme. This fix sets the minimum size of a
button to 0x0 so that it can take on the image size regardless of what
the platform's default dimensions for a button.

Change-Id: I590eb1bec860cda221e821c663c03c5cd81a65fe